+++ scripts/SoundManager.cs
@@ -370,8 +370,8 @@ public partial class SoundManager : Node, ISkinnable
 
     public static float ComputeVolumeDb(float volume, float master, float range)
     {
-        if (volume <= 0 || master <= 0) return float.NegativeInfinity;
-        return -80 + range * (float)Math.Pow(volume / 100, 0.1) * (float)Math.Pow(master / 100, 0.1);
+        if (volume <= 0 || master <= 0) return double.NegativeInfinity;
+        return -80 + range * Math.Pow(volume / 100, 0.1) * Math.Pow(master / 100, 0.1);
     }
 
     public static void UpdateVolume()
